requirements.txt
matplotlib
numpy
opencv-python
Pillow
PyYAML
requests
scipy
torch
torchvision
tqdm
tensorboard
pandas
seaborn
argparse
pathlib
streamlit
pytest-shutil

